feature(bento-pan-zoom): organized hooks, added doc comments
